+++ js/index.js
@@ -1,4 +1,5 @@
 import { api, greenColor, redColor, secondaryLabel } from "./constants.js"
+import { updateChartContainer } from "./chart.js"
 
 const tickerIndexEls = document.getElementsByClassName("tickerIndex")
 const companyNameEls = document.getElementsByClassName("companyName")
@@ -96,78 +97,6 @@ function updateStatsContainer({ range: { high, low }, open, volume, avg_volume, 
   }
 }
 
-function updateStockGraph(data) {
-  console.log(data)
-
-  let dates = []
-  let points = []
-  for (let point of data) {
-    dates.push(point.date.slice(6))
-    points.push(point.close)
-  }
-
-  let lineColor = points[0] < points[points.length - 1] ? greenColor : redColor
-  let stepSize = Math.abs(points[0] - points[points.length - 1])
-
-  let gridStyles = {
-    drawBorder: false,
-    drawOnChartArea: false,
-    drawTicks: false,
-  }
-
-  var ctx = document.getElementById("myChart").getContext("2d")
-  var myChart = new Chart(ctx, {
-    type: "line",
-    data: {
-      labels: dates,
-      datasets: [
-        {
-          label: "Price",
-          data: points,
-          borderColor: [lineColor],
-          pointBorderColor: lineColor,
-          borderWidth: 5,
-          fill: false,
-          lineTension: 0,
-          responsive: true,
-          maintainAspectRatio: false,
-        },
-      ],
-    },
-    options: {
-      legend: {
-        display: false,
-      },
-      animation: {
-        duration: 0,
-      },
-      scales: {
-        yAxes: [
-          {
-            ticks: {
-              fontColor: secondaryLabel,
-              fontSize: 16,
-              stepSize: 4,
-              padding: 10,
-            },
-            gridLines: gridStyles,
-          },
-        ],
-        xAxes: [
-          {
-            ticks: {
-              fontColor: secondaryLabel,
-              fontSize: 16,
-              padding: 10,
-            },
-            gridLines: gridStyles,
-          },
-        ],
-      },
-    },
-  })
-}
-
 function setLoadingState(isLoading) {
   searchButton.disabled = isLoading
   searchInput.disabled = isLoading
@@ -218,8 +147,8 @@ async function requestData(ticker) {
       updateCompanyContainer(result)
       updatePriceContainer(result)
       updateStatsContainer(result)
+      updateChartContainer(result.historical)
       updateTabTitle(result)
-      updateStockGraph(result.historical)
     })
 
     .catch((err) => {
